+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - Changed price value restriction
 - Added global exception handler
 - Added shopware link support
-- Added customer order tax free support 
+- Added customer order tax free support
+- Added specific ci encoding support
 
 1.0.11
 ------